Simplified Kakoune comms with kak-tree-sitter server.
Initially, Kakoune would have to send a parse and then a highlight
command to the kak-tree-sitter server in order for highlighting to
happen. This not only overcomplicated how highlighting is handled,
but also created a bug where if the two commands were sent too
quickly, kak-tree-sitter would crash due to not being able to parse
both messages at once. This commit fixes these issues.