Cleaned up and optimized comms with kakoune-treesitter server.
To be honest, it's starting to look like Python was the wrong pick
for this plugin. Its startup time is just too slow for something that
needs to be ready to go right when a buffer is opened.

From now on, I'm going to switch to Go as an implementation language.
It's a lot simpler than Rust, but has a substantial performance
advantage over Python.